refactor(ingredient): co-locate measurement tests with their submodules
measurement/mod.rs was ~76% tests, exercising functions defined in the
sibling submodules (which had zero internal tests). Move each test into the
submodule that owns the function under test (single/range/composite/number/
guards); mod.rs keeps only the parse_measurement_list / separator / dispatch
tests. A shared units() fixture now lives in a #[cfg(test)] test_support
module, delegated to by a local #[fixture] in each submodule. Pure test
relocation — 688 tests, same coverage.
